Check if any votes exist, then delete

+++ src/app/api/common/comments/deleteImpactMetricComment.ts
@@ -7,22 +7,47 @@ async function deleteImpactMetricCommentApi({
   commentId: number;
 }) {
 
-
-  const deletedVotes = await prisma.metrics_comments_votes.deleteMany({
-    where: {
-      comment_id: commentId
-    }
-  });
-
-  const deletedComment = await prisma.metrics_comments.delete({
+  const votesExist = await prisma.metrics_comments_votes.findMany({
     where: {
       comment_id: commentId,
-    }
+    },
   });
 
-  return {
-    comment_id: deletedComment.comment_id,
-  };
+  // If related votes exist, delete them
+  if (votesExist.length > 0) {
+    await prisma.metrics_comments_votes.deleteMany({
+      where: {
+        comment_id: commentId,
+      },
+    });
+  }
+
+  // Check if the comment exists before attempting to delete it
+  const commentExists = await prisma.metrics_comments.findUnique({
+    where: {
+      comment_id: commentId,
+    },
+  });
+
+  if (commentExists) {
+    // Delete the record in the main table
+    const deletedComment = await prisma.metrics_comments.delete({
+      where: {
+        comment_id: commentId,
+      },
+    });
+
+    return {
+      comment_id: deletedComment.comment_id,
+    };
+  }
+
+  else {
+    return {
+      comment_id: -1,
+    }
+  }
+
 }
 
 export const deleteImpactMetricComment = cache(deleteImpactMetricCommentApi);
